fix: run contract review synchronously and various bug fixes
- Replace asyncio.create_task with synchronous review to avoid background
  task being killed on server reload
- Fix review_reports UNIQUE constraint conflict by deleting existing
  data before re-inserting
- Fix PendingRollbackError in exception handler (rollback before retry)
- Add 120s timeout to DeepSeek API client
- Add new admin models/schemas, Review.vue page, upload store
- Remove deprecated templates module
- Add root .gitignore excluding uploads/ and venv/
- Update README with demo video

+++ backend/app/modules/risks/schemas.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 class RiskItemResponse(BaseModel):
     id: int
     contract_id: int
-    clause_index: int
+    clause_index: float
     clause_text: str
     risk_level: str
     category: str
